+++ plot_data/core.py
@@ -15,7 +15,6 @@ from typing import Any, Dict, List, Tuple, Union
 
 import matplotlib.pyplot as plt
 from matplotlib.patches import Polygon
-import numpy as npy
 
 try:
     # dessia_common >= 0.12.0
@@ -32,8 +31,6 @@ from matplotlib import patches
 
 import plot_data.colors
 from plot_data import templates
-
-npy.seterr(divide='raise')
 
 
 def delete_none_from_dict(dict1):
@@ -759,6 +756,7 @@ class Dataset(PlotDataObject):
             else:
                 raise ValueError(f"Element of type {type(element)} cannot be used as a Dataset data element.")
         self.elements = sampled_elements
+        self.partial_points = partial_points
         self.display_step = display_step
         PlotDataObject.__init__(self, type_='dataset', name=name)