+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.14.0]
+### Add
+- to_html_file method in Python so that it is possible to create a plot data file without opening it in web browser
+
 ### Fix
 - Python 3.9 in drone