+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fix
 - Smart writing of axes' names in ParallelPlot
+- Allow to draw all points of a Dataset in Graph objects
 - Python 3.9 in drone
+- Remove numpy from dependencies
 
 ## [0.13.0]
 ### Add