+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.14.0]
 ### Add
+- to_html method in Python so that it is possible to create a plot data file without opening it in web browser
+- Allow to draw all points of a Dataset in Graph objects
 
 ### Fix
 - Smart writing of axes' names in ParallelPlot