+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0]
+### Add
+- to_html method in Python so that it is possible to create a plot data file without opening it in web browser
+- Allow to draw all points of a Dataset in Graph objects
+
+### Fix
+- Smart writing of axes' names in ParallelPlot
+- Allow to draw all points of a Dataset in Graph objects
+- Python 3.9 in drone
+- Remove numpy from dependencies
+
 ## [0.13.0]
 ### Add
 - Doc Typescript